+Update
Died due to a hanging session with the included farmer.
Many things that don't suit me (cesspool in the code, lack of dependency).
Stop it.
It's time to tackle a new client. Most likely I will not do new things here (only workability). Let the client die in peace.
People who wish to participate in the creation of a new era are welcome. (RUS)